Initialize numerous variables
A bunch of buffers in the MusicLeds class were left unintialized, which
causes problems when you're using dynamically allocated memory.

Ran the default initializer on all of the buffers, which made the class
behave properly when initialized as an object instance off of the heap,
instead of being a statically declared stack object.